+++ Gemfile
@@ -56,5 +56,6 @@ group :development, :test do
   # Use database_cleaner in place of transactional fixtures
   gem 'database_cleaner'
   # Use pry-rails for debugging
-  gem 'pry-rails', '~> 0.3.4'
+  gem 'pry-rails'
+  gem "pry-byebug"
 end
