+++ app/controllers/users_controller.rb
@@ -1,5 +1,10 @@
 class UsersController < ApplicationController
   def show
+    if params[:id]
+      @user = User.find(params[:id])
+    else
+      @user = current_user
+    end
   end
 
   def edit
