+++ app/views/layouts/application.html.erb
@@ -40,10 +40,14 @@
     </div>
   </nav>
 
-  <p class="notice"><%= notice %></p>
-  <p class="alert"><%= alert %></p>
+  <div class="container">
+    <div class="row">
+      <p class="notice"><%= notice %></p>
+      <p class="alert"><%= alert %></p>
 
-  <%= yield %>
+      <%= yield %>
+    </div>
+  </div>
 
 </body>
 </html>
