+++ app/views/users/show.html.erb
@@ -1,2 +1,9 @@
-<h1>Users#show</h1>
-<p>Find me in app/views/users/show.html.erb</p>
+<div class="container">
+  <div class="row">
+    <div class="col-xs-12">
+      <h1 id="username"><%= @user.username %> Profile</h1>
+      <p id="email"><%= @user.email %></p>
+      <p id="biography"><%= @user.biography %></p>
+    </div>
+  </div
+</div>
