+++ spec/factories/users.rb
@@ -2,7 +2,7 @@ FactoryGirl.define do
   factory :user do
     email { Faker::Internet.email }
     password { Faker::Internet.password }
-    username { Faker::Internet.name }
-    biography { Faker::Lorem.paragraph(3) }
+    username { Faker::Internet.user_name }
+    biography { Faker::Lorem.paragraphs(3) }
   end
 end
