+++ spec/rails_helper.rb
@@ -7,6 +7,8 @@ require 'spec_helper'
 require 'rspec/rails'
 require 'capybara/rails'
 
+require "support/session_helpers"
+
 # Add additional requires below this line. Rails is not loaded until this point!
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
@@ -82,4 +84,7 @@ RSpec.configure do |config|
 
   # include Devise test helpers
   config.include Devise::TestHelpers, type: :controller
+
+  # include our SessionHelpers to signup/login/logout
+  config.include SessionHelpers, type: :feature
 end
